Correction/changes to function strucfunc_plot() in bplot.py
1. Correction: make sure to use ~to_fit instead of large_scale when
plotting the points that are omitted from the fit, since they are not
always the same

2. Change of behavior: pass in result_orig, so we can use the best-fit
parameters from the original lev-marq fit.  This is for consistency
with the corner plots, where we plot these values (truth) in blue

HOWEVER, I am not convinced that this is the best thing to do. There
are some pathological cases where the lev-marq best fit is not well
centered on the emcee confidence basin.  But these are cases where the
posterior distributions are bimodal or worse.  Maybe we should find
the maximum-likelihood values from the emcee fit (the values that we
previously used are just the median over the walker chains)