Improvements to the Lagoon fitting
We can now get a reasonable fit with good error bounds, even when
using the full map. The main thing that I had to change to achieve
this was to use more reasonable bounds on the parameters.  What Javier
had was not allowing the sig2 to go low enough because the bounds were
based on the full structure function range including the large-scale
parts that are omitted from the fit.

I also doubled the emcee steps to 10000, so that it gets an estimate
for the autocorrelation times.

Finally, I made use of the new bplot.STYLE to improve the placement of
the labels